- Implemented createSchemaUtils - Began testing the schema-based utils
