+++ packages/utils/src/index.ts
@@ -60,6 +60,7 @@ import {
   RegistryWidgetsType,
   RJSFSchema,
   RJSFValidationError,
+  SchemaUtilsType,
   TitleFieldProps,
   UiSchema,
   UIOptionsType,
@@ -100,6 +101,7 @@ export type {
   RegistryWidgetsType,
   RJSFSchema,
   RJSFValidationError,
+  SchemaUtilsType,
   TitleFieldProps,
   UiSchema,
   UIOptionsType,
