+++ packages/utils/test/schema/utils.test.js
@@ -25,1120 +25,6 @@ describe('utils', () => {
     sandbox.restore();
   });
 
-  describe('getDefaultFormState()', () => {
-    describe('root default', () => {
-      it('should map root schema default to form state, if any', () => {
-        expect(
-          getDefaultFormState({
-            type: 'string',
-            default: 'foo',
-          })
-        ).to.eql('foo');
-      });
-
-      it('should keep existing form data that is equal to 0', () => {
-        expect(
-          getDefaultFormState(
-            {
-              type: 'number',
-              default: 1,
-            },
-            0
-          )
-        ).to.eql(0);
-      });
-
-      it('should keep existing form data that is equal to false', () => {
-        expect(
-          getDefaultFormState(
-            {
-              type: 'boolean',
-            },
-            false
-          )
-        ).to.eql(false);
-      });
-
-      const noneValues = [null, undefined, NaN];
-      noneValues.forEach(noneValue => {
-        it('should overwrite existing form data that is equal to a none value', () => {
-          expect(
-            getDefaultFormState(
-              {
-                type: 'number',
-                default: 1,
-              },
-              noneValue
-            )
-          ).to.eql(1);
-        });
-      });
-    });
-
-    describe('nested default', () => {
-      it('should map schema object prop default to form state', () => {
-        expect(
-          getDefaultFormState({
-            type: 'object',
-            properties: {
-              string: {
-                type: 'string',
-                default: 'foo',
-              },
-            },
-          })
-        ).to.eql({ string: 'foo' });
-      });
-
-      it('should default to empty object if no properties are defined', () => {
-        expect(
-          getDefaultFormState({
-            type: 'object',
-          })
-        ).to.eql({});
-      });
-
-      it('should recursively map schema object default to form state', () => {
-        expect(
-          getDefaultFormState({
-            type: 'object',
-            properties: {
-              object: {
-                type: 'object',
-                properties: {
-                  string: {
-                    type: 'string',
-                    default: 'foo',
-                  },
-                },
-              },
-            },
-          })
-        ).to.eql({ object: { string: 'foo' } });
-      });
-
-      it('should map schema array default to form state', () => {
-        expect(
-          getDefaultFormState({
-            type: 'object',
-            properties: {
-              array: {
-                type: 'array',
-                default: ['foo', 'bar'],
-                items: {
-                  type: 'string',
-                },
-              },
-            },
-          })
-        ).to.eql({ array: ['foo', 'bar'] });
-      });
-
-      it('should recursively map schema array default to form state', () => {
-        expect(
-          getDefaultFormState({
-            type: 'object',
-            properties: {
-              object: {
-                type: 'object',
-                properties: {
-                  array: {
-                    type: 'array',
-                    default: ['foo', 'bar'],
-                    items: {
-                      type: 'string',
-                    },
-                  },
-                },
-              },
-            },
-          })
-        ).to.eql({ object: { array: ['foo', 'bar'] } });
-      });
-
-      it('should propagate nested defaults to resulting formData by default', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            object: {
-              type: 'object',
-              properties: {
-                array: {
-                  type: 'array',
-                  default: ['foo', 'bar'],
-                  items: {
-                    type: 'string',
-                  },
-                },
-                bool: {
-                  type: 'boolean',
-                  default: true,
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          object: { array: ['foo', 'bar'], bool: true },
-        });
-      });
-
-      it('should keep parent defaults if they don`t have a node level default', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            level1: {
-              type: 'object',
-              default: {
-                level2: {
-                  leaf1: 1,
-                  leaf2: 1,
-                  leaf3: 1,
-                  leaf4: 1,
-                },
-              },
-              properties: {
-                level2: {
-                  type: 'object',
-                  default: {
-                    // No level2 default for leaf1
-                    leaf2: 2,
-                    leaf3: 2,
-                  },
-                  properties: {
-                    leaf1: { type: 'number' }, // No level2 default for leaf1
-                    leaf2: { type: 'number' }, // No level3 default for leaf2
-                    leaf3: { type: 'number', default: 3 },
-                    leaf4: { type: 'number' }, // Defined in formData.
-                  },
-                },
-              },
-            },
-          },
-        };
-        expect(
-          getDefaultFormState(schema, {
-            level1: { level2: { leaf4: 4 } },
-          })
-        ).eql({
-          level1: {
-            level2: { leaf1: 1, leaf2: 2, leaf3: 3, leaf4: 4 },
-          },
-        });
-      });
-
-      it('should support nested values in formData', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            level1: {
-              type: 'object',
-              properties: {
-                level2: {
-                  oneOf: [
-                    {
-                      type: 'object',
-                      properties: {
-                        leaf1: {
-                          type: 'string',
-                        },
-                      },
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        };
-        const formData = {
-          level1: {
-            level2: {
-              leaf1: 'a',
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, formData)).eql({
-          level1: { level2: { leaf1: 'a' } },
-        });
-      });
-
-      it('should use parent defaults for ArrayFields', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            level1: {
-              type: 'array',
-              default: [1, 2, 3],
-              items: { type: 'number' },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: [1, 2, 3],
-        });
-      });
-
-      it('should use parent defaults for ArrayFields if declared in parent', () => {
-        const schema = {
-          type: 'object',
-          default: { level1: [1, 2, 3] },
-          properties: {
-            level1: {
-              type: 'array',
-              items: { type: 'number' },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: [1, 2, 3],
-        });
-      });
-
-      it('should map item defaults to fixed array default', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            array: {
-              type: 'array',
-              items: [
-                {
-                  type: 'string',
-                  default: 'foo',
-                },
-                {
-                  type: 'number',
-                },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          array: ['foo', undefined],
-        });
-      });
-
-      it('should merge schema array item defaults from grandparent for overlapping default definitions', () => {
-        const schema = {
-          type: 'object',
-          default: {
-            level1: { level2: ['root-default-1', 'root-default-2'] },
-          },
-          properties: {
-            level1: {
-              type: 'object',
-              properties: {
-                level2: {
-                  type: 'array',
-                  items: [
-                    {
-                      type: 'string',
-                      default: 'child-default-1',
-                    },
-                    {
-                      type: 'string',
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        };
-
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: { level2: ['child-default-1', 'root-default-2'] },
-        });
-      });
-
-      it('should overwrite schema array item defaults from parent for nested default definitions', () => {
-        const schema = {
-          type: 'object',
-          default: {
-            level1: {
-              level2: [{ item: 'root-default-1' }, { item: 'root-default-2' }],
-            },
-          },
-          properties: {
-            level1: {
-              type: 'object',
-              default: { level2: [{ item: 'parent-default-1' }, {}] },
-              properties: {
-                level2: {
-                  type: 'array',
-                  items: {
-                    type: 'object',
-                    properties: {
-                      item: {
-                        type: 'string',
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        };
-
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: { level2: [{ item: 'parent-default-1' }, {}] },
-        });
-      });
-
-      it('should merge schema array item defaults from the same item for overlapping default definitions', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            level1: {
-              type: 'array',
-              default: ['property-default-1', 'property-default-2'],
-              items: [
-                {
-                  type: 'string',
-                  default: 'child-default-1',
-                },
-                {
-                  type: 'string',
-                },
-              ],
-            },
-          },
-        };
-
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: ['child-default-1', 'property-default-2'],
-        });
-      });
-
-      it('should merge schema from additionalItems defaults into property default', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            level1: {
-              type: 'array',
-              default: [
-                {
-                  item: 'property-default-1',
-                },
-                {},
-              ],
-              additionalItems: {
-                type: 'object',
-                properties: {
-                  item: {
-                    type: 'string',
-                    default: 'additional-default',
-                  },
-                },
-              },
-              items: [
-                {
-                  type: 'object',
-                  properties: {
-                    item: {
-                      type: 'string',
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: [
-            { item: 'property-default-1' },
-            { item: 'additional-default' },
-          ],
-        });
-      });
-
-      it('should overwrite defaults over multiple levels with arrays', () => {
-        const schema = {
-          type: 'object',
-          default: {
-            level1: [
-              {
-                item: 'root-default-1',
-              },
-              {
-                item: 'root-default-2',
-              },
-              {
-                item: 'root-default-3',
-              },
-              {
-                item: 'root-default-4',
-              },
-            ],
-          },
-          properties: {
-            level1: {
-              type: 'array',
-              default: [
-                {
-                  item: 'property-default-1',
-                },
-                {},
-                {},
-              ],
-              additionalItems: {
-                type: 'object',
-                properties: {
-                  item: {
-                    type: 'string',
-                    default: 'additional-default',
-                  },
-                },
-              },
-              items: [
-                {
-                  type: 'object',
-                  properties: {
-                    item: {
-                      type: 'string',
-                    },
-                  },
-                },
-                {
-                  type: 'object',
-                  properties: {
-                    item: {
-                      type: 'string',
-                      default: 'child-default-2',
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-
-        expect(getDefaultFormState(schema, {})).eql({
-          level1: [
-            { item: 'property-default-1' },
-            { item: 'child-default-2' },
-            { item: 'additional-default' },
-          ],
-        });
-      });
-
-      it('should use schema default for referenced definitions', () => {
-        const schema = {
-          definitions: {
-            testdef: {
-              type: 'object',
-              properties: {
-                foo: { type: 'number' },
-              },
-            },
-          },
-          $ref: '#/definitions/testdef',
-          default: { foo: 42 },
-        };
-
-        expect(getDefaultFormState(schema, undefined, schema)).eql({
-          foo: 42,
-        });
-      });
-
-      it('should fill array with additional items schema when items is empty', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            array: {
-              type: 'array',
-              minItems: 1,
-              additionalItems: {
-                type: 'string',
-                default: 'foo',
-              },
-              items: [],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          array: ['foo'],
-        });
-      });
-
-      it('should not fill array with additional items from schema when items is empty and form data contains partial array', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            array: {
-              type: 'array',
-              minItems: 2,
-              additionalItems: {
-                type: 'string',
-                default: 'foo',
-              },
-              items: [],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { array: ['bar'] })).eql({
-          array: ['bar'],
-        });
-      });
-
-      it('should fill defaults in existing array items', () => {
-        const schema = {
-          type: 'array',
-          minItems: 2,
-          items: {
-            type: 'object',
-            properties: {
-              item: {
-                type: 'string',
-                default: 'foo',
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, [{}])).eql([{ item: 'foo' }]);
-      });
-
-      it('defaults passed along for multiselect arrays when minItems is present', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            array: {
-              type: 'array',
-              minItems: 1,
-              uniqueItems: true,
-              default: ['foo', 'qux'],
-              items: {
-                type: 'string',
-                enum: ['foo', 'bar', 'fuzz', 'qux'],
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          array: ['foo', 'qux'],
-        });
-      });
-    });
-
-    describe('defaults with oneOf', () => {
-      it('should populate defaults for oneOf', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'string',
-              oneOf: [
-                { type: 'string', default: 'a' },
-                { type: 'string', default: 'b' },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          name: 'a',
-        });
-      });
-
-      it('should populate defaults for oneOf when `type`: `object` is missing', () => {
-        const schema = {
-          type: 'object',
-          oneOf: [
-            {
-              properties: { name: { type: 'string', default: 'a' } },
-            },
-            {
-              properties: { id: { type: 'number', default: 13 } },
-            },
-          ],
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          name: 'a',
-        });
-      });
-
-      it('should populate nested default values for oneOf', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'object',
-              oneOf: [
-                {
-                  type: 'object',
-                  properties: {
-                    first: { type: 'string', default: 'First Name' },
-                  },
-                },
-                { type: 'string', default: 'b' },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          name: {
-            first: 'First Name',
-          },
-        });
-      });
-
-      it('should populate defaults for oneOf + dependencies', () => {
-        const schema = {
-          oneOf: [
-            {
-              type: 'object',
-              properties: {
-                name: {
-                  type: 'string',
-                },
-              },
-            },
-          ],
-          dependencies: {
-            name: {
-              oneOf: [
-                {
-                  properties: {
-                    name: {
-                      type: 'string',
-                    },
-                    grade: {
-                      default: 'A',
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { name: 'Name' })).eql({
-          name: 'Name',
-          grade: 'A',
-        });
-      });
-    });
-
-    describe('defaults with anyOf', () => {
-      it('should populate defaults for anyOf', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'string',
-              anyOf: [
-                { type: 'string', default: 'a' },
-                { type: 'string', default: 'b' },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          name: 'a',
-        });
-      });
-
-      it('should populate nested default values for anyOf', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'object',
-              anyOf: [
-                {
-                  type: 'object',
-                  properties: {
-                    first: { type: 'string', default: 'First Name' },
-                  },
-                },
-                { type: 'string', default: 'b' },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, {})).eql({
-          name: {
-            first: 'First Name',
-          },
-        });
-      });
-
-      it('should populate defaults for anyOf + dependencies', () => {
-        const schema = {
-          anyOf: [
-            {
-              type: 'object',
-              properties: {
-                name: {
-                  type: 'string',
-                },
-              },
-            },
-          ],
-          dependencies: {
-            name: {
-              oneOf: [
-                {
-                  properties: {
-                    name: {
-                      type: 'string',
-                    },
-                    grade: {
-                      type: 'string',
-                      default: 'A',
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { name: 'Name' })).eql({
-          name: 'Name',
-          grade: 'A',
-        });
-      });
-    });
-
-    describe('with dependencies', () => {
-      it('should populate defaults for dependencies', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'string',
-            },
-          },
-          dependencies: {
-            name: {
-              oneOf: [
-                {
-                  properties: {
-                    name: {
-                      type: 'string',
-                    },
-                    grade: {
-                      type: 'string',
-                      default: 'A',
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { name: 'Name' })).eql({
-          name: 'Name',
-          grade: 'A',
-        });
-      });
-
-      it('should populate defaults for nested dependencies', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            foo: {
-              type: 'object',
-              properties: {
-                name: {
-                  type: 'string',
-                },
-              },
-              dependencies: {
-                name: {
-                  oneOf: [
-                    {
-                      properties: {
-                        name: {
-                          type: 'string',
-                        },
-                        grade: {
-                          type: 'string',
-                          default: 'A',
-                        },
-                      },
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { foo: { name: 'Name' } })).eql({
-          foo: {
-            name: 'Name',
-            grade: 'A',
-          },
-        });
-      });
-
-      it('should populate defaults for nested dependencies in arrays', () => {
-        const schema = {
-          type: 'array',
-          items: {
-            properties: {
-              foo: {
-                type: 'object',
-                properties: {
-                  name: {
-                    type: 'string',
-                  },
-                },
-                dependencies: {
-                  name: {
-                    oneOf: [
-                      {
-                        properties: {
-                          name: {
-                            type: 'string',
-                          },
-                          grade: {
-                            type: 'string',
-                            default: 'A',
-                          },
-                        },
-                      },
-                    ],
-                  },
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, [{ foo: { name: 'Name' } }])).eql([
-          {
-            foo: {
-              name: 'Name',
-              grade: 'A',
-            },
-          },
-        ]);
-      });
-
-      it('should populate defaults for nested dependencies in arrays when matching enum values in oneOf', () => {
-        const schema = {
-          type: 'array',
-          items: {
-            properties: {
-              foo: {
-                type: 'object',
-                properties: {
-                  name: {
-                    type: 'string',
-                  },
-                },
-                dependencies: {
-                  name: {
-                    oneOf: [
-                      {
-                        properties: {
-                          name: {
-                            enum: ['first'],
-                          },
-                          grade: {
-                            type: 'string',
-                            default: 'A',
-                          },
-                        },
-                      },
-                      {
-                        properties: {
-                          name: {
-                            enum: ['second'],
-                          },
-                          grade: {
-                            type: 'string',
-                            default: 'B',
-                          },
-                        },
-                      },
-                    ],
-                  },
-                },
-              },
-            },
-          },
-        };
-        expect(
-          getDefaultFormState(schema, [
-            { foo: { name: 'first' } },
-            { foo: { name: 'second' } },
-            { foo: { name: 'third' } },
-          ])
-        ).eql([
-          {
-            foo: {
-              name: 'first',
-              grade: 'A',
-            },
-          },
-          {
-            foo: {
-              name: 'second',
-              grade: 'B',
-            },
-          },
-          {
-            foo: {
-              name: 'third',
-            },
-          },
-        ]);
-      });
-
-      it('should populate defaults for nested oneOf + dependencies', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            foo: {
-              oneOf: [
-                {
-                  type: 'object',
-                  properties: {
-                    name: {
-                      type: 'string',
-                    },
-                  },
-                },
-              ],
-              dependencies: {
-                name: {
-                  oneOf: [
-                    {
-                      properties: {
-                        name: {
-                          type: 'string',
-                        },
-                        grade: {
-                          type: 'string',
-                          default: 'A',
-                        },
-                      },
-                    },
-                  ],
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { foo: { name: 'Name' } })).eql({
-          foo: {
-            name: 'Name',
-            grade: 'A',
-          },
-        });
-      });
-
-      it('should populate defaults for nested dependencies when formData passed to computeDefaults is undefined', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            can_1: {
-              type: 'object',
-              properties: {
-                phy: {
-                  title: 'Physical',
-                  description: 'XYZ',
-                  type: 'object',
-                  properties: {
-                    bit_rate_cfg_mode: {
-                      title: 'Sub title',
-                      description: 'XYZ',
-                      type: 'integer',
-                      default: 0,
-                    },
-                  },
-                  dependencies: {
-                    bit_rate_cfg_mode: {
-                      oneOf: [
-                        {
-                          properties: {
-                            bit_rate_cfg_mode: {
-                              enum: [0],
-                            },
-                          },
-                        },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, undefined)).eql({
-          can_1: {
-            phy: {
-              bit_rate_cfg_mode: 0,
-            },
-          },
-        });
-      });
-
-      it('should not crash for defaults for nested dependencies when formData passed to computeDefaults is null', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            can_1: {
-              type: 'object',
-              properties: {
-                phy: {
-                  title: 'Physical',
-                  description: 'XYZ',
-                  type: 'object',
-                  properties: {
-                    bit_rate_cfg_mode: {
-                      title: 'Sub title',
-                      description: 'XYZ',
-                      type: 'integer',
-                      default: 0,
-                    },
-                  },
-                  dependencies: {
-                    bit_rate_cfg_mode: {
-                      oneOf: [
-                        {
-                          properties: {
-                            bit_rate_cfg_mode: {
-                              enum: [0],
-                            },
-                          },
-                        },
-                      ],
-                    },
-                  },
-                },
-              },
-            },
-          },
-        };
-        expect(getDefaultFormState(schema, { can_1: { phy: null } })).eql({
-          can_1: {
-            phy: null,
-          },
-        });
-      });
-    });
-
-    describe('with schema keys not defined in the formData', () => {
-      it('shouldn`t add in undefined keys to formData', () => {
-        const schema = {
-          type: 'object',
-          properties: {
-            foo: { type: 'string' },
-            bar: { type: 'string' },
-          },
-        };
-        const formData = {
-          foo: 'foo',
-          baz: 'baz',
-        };
-        const result = {
-          foo: 'foo',
-          baz: 'baz',
-        };
-        expect(getDefaultFormState(schema, formData)).to.eql(result);
-      });
-    });
-  });
-
   describe('isMultiSelect()', () => {
     describe('uniqueItems is true', () => {
       describe('schema items enum is an array', () => {
